Set up Vite React app for simulator
